backend/requirements.txt
ppdirs==1.4.4
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
dnspython==2.4.2
frozendict==2.3.8
html5lib==1.1
idna==3.4
linecache2==1.0.0
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.0
pandas==2.1.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
traceback2==1.4.0
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.6
webencodings==0.5.1
yfinance==0.2.28
bs4==0.0.1